+++ static/components/f-code.js
@@ -1,4 +1,4 @@
-import { asset, oninvalid, updateField, view } from "./utils.js";
+import { asset, oninvalid, pushOptions, updateField, view } from "./utils.js";
 import { Component } from "./component.js";
 import { init } from "../libs/code.js";
 import dom from "dom";
@@ -23,6 +23,8 @@ customElements.define(
         html: schema.description,
       }, this);
 
+      const tools = dom("div", { class: "tools" }, this);
+
       const textarea = dom("textarea", {
         id,
         name,
@@ -38,7 +40,17 @@ customElements.define(
       `;
 
       const code = dom("div", { class: "code" }, shadow);
-      this.editor = init(code, textarea).editor;
+      const { editor, changeLanguage, allLanguages } = init(code, textarea);
+      this.editor = editor;
+      const languageSelector = dom("select", {
+        class: "select",
+        onchange: (event) => {
+          changeLanguage(event.target.value);
+        },
+      }, tools);
+      pushOptions(languageSelector, allLanguages);
+      languageSelector.value = schema.attributes?.data?.language ?? "HTML";
+      languageSelector.dispatchEvent(new Event("change"));
     }
 
     get currentValue() {